[EA-2025-08-31] 完整修復包：Buildings/Core/Creep 系統整合
包含內容：
- Buildings: ResearchService 介面修復、ResourceType 枚舉、兼容性層
- Core: GameBootstrapper 更新、工具類（ArrayUtils, CoroutineStub）
- Creep: Manager 更新、CreepSourceTypeCompat 支援
- Units: UnitManager 服務整合
- Utils: ResourceCostShim 兼容性工具

修復編譯錯誤並提供向後兼容性支援
EA-2025-08-31-FULL

+++ Assets/DeepAbyssHive/Creep/Managers/CreepManager.Query.cs
@@ -321,15 +321,15 @@ namespace DeepAbyssHive.Creep.Managers
                 stats.TotalHealth += tile.Health;
                 stats.TotalResourcesGenerated += tile.TotalResourcesGenerated;
                 
-                switch (tile.Status)
+                switch ((int)tile.Status)
                 {
-                    case DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Healthy:
+                    case (int)DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Healthy:
                         stats.HealthyTiles++;
                         break;
-                    case DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Weakened:
+                    case (int)DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Weakened:
                         stats.GrowingTiles++;
                         break;
-                    case DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Collapsing:
+                    case (int)DeepAbyssHive.Creep.Compat.CreepTileStatusCompat.Collapsing:
                         stats.DyingTiles++;
                         break;
                 }
